refactor: moved dataset dtos into separate folder and changed namespace
