+++ src/Sepes.Infrastructure/Dto/Study/StudyDetailsDto.cs
@@ -1,4 +1,5 @@
-﻿using Sepes.Infrastructure.Dto.Interfaces;
+﻿using Sepes.Infrastructure.Dto.Dataset;
+using Sepes.Infrastructure.Dto.Interfaces;
 using Sepes.Infrastructure.Dto.Sandbox;
 using Sepes.Infrastructure.Interface;
 using System.Collections.Generic;
